pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
typing-extensions = "^4.1.1"
yarl = "^1.7.2"
aiohttp = "^3.8.1"
orjson = "^3.6.7"
requests = "^2.27.1"
Flask = "^2.0.3"
lavalink = "^3.1.5"
youtube_dl = "^2021.12.17"
discord-pretty-help = "^1.3.3"
bs4 = "^0.0.1"
PyNaCl = "^1.5.0"
pymongo = "^4.0.2"
dnspython = "^2.2.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"